+++ src/main/java/com/altima/springboot/app/models/entity/AmpTraspasoDetalle.java
@@ -36,7 +36,7 @@ public class AmpTraspasoDetalle implements Serializable{
 	private String tipo;
 	
 	@Column(name="cantidad")
-	private Long cantidad;
+	private Float cantidad;
 	
 	@Column(name="descuento_mon")
 	private String descuentoMon;
@@ -79,11 +79,11 @@ public class AmpTraspasoDetalle implements Serializable{
 		this.tipo = tipo;
 	}
 
-	public Long getCantidad() {
+	public Float getCantidad() {
 		return cantidad;
 	}
 
-	public void setCantidad(Long cantidad) {
+	public void setCantidad(Float cantidad) {
 		this.cantidad = cantidad;
 	}
